Using new nullability annonations to correctly tell Swift messages might be nil.
The Swift prototype for listenForMessageWithIdentifier was:

    func listenForMessageWithIdentifier(identifier: String!, listener: ((AnyObject!) -> Void)!)

Thing is, that there are race conditions that lead to MMWormhole calling the listener with a nill message. The annonations change that prototype to:

func listenForMessageWithIdentifier(identifier: String, listener: ((AnyObject?) -> Void)?)

This means the listener block has to unwrap the message and won't crash on nil.